+++ tests/test_visitors/test_tokenize/test_wrong_primitives/test_partial_floats.py
@@ -10,42 +10,49 @@ from wemake_python_styleguide.visitors.tokenize.primitives import (
 )
 
 
-@pytest.mark.parametrize('code', [
-    'x = 10.',
-    'print(3. + 5.1)',
-    '.05 + 1.2',
-    'words: int = 12.',
+@pytest.mark.parametrize('primitive', [
+    '10.',
+    '.05',
+    '.0',
+    '0.',
+    '-5.',
+    '-.43',
 ])
 def test_partial_float(
     parse_tokens,
     assert_errors,
+    assert_error_text,
     default_options,
-    code,
+    primitives_usages,
+    primitive,
+    mode,
 ):
     """Ensures that partial floats raise a warning."""
-    file_tokens = parse_tokens(code)
+    file_tokens = parse_tokens(mode(primitives_usages.format(primitive)))
 
     visitor = WrongPrimitivesVisitor(default_options, file_tokens=file_tokens)
     visitor.run()
 
     assert_errors(visitor, [PartialFloatViolation])
+    assert_error_text(visitor, primitive.replace('-', ''))
 
 
-@pytest.mark.parametrize('code', [
-    'x = 10.0',
-    'print(3.0 + 5.1)',
-    '0.05 + 1.2',
-    'words: int = 12',
-    '# comment: .2',
+@pytest.mark.parametrize('primitive', [
+    '10.0',
+    '0.14',
+    '-0.05',
+    '-1.1',
 ])
 def test_correct_float(
     parse_tokens,
     assert_errors,
     default_options,
-    code,
+    primitives_usages,
+    primitive,
+    mode,
 ):
     """Ensures that correct floats are fine."""
-    file_tokens = parse_tokens(code)
+    file_tokens = parse_tokens(mode(primitives_usages.format(primitive)))
 
     visitor = WrongPrimitivesVisitor(default_options, file_tokens=file_tokens)
     visitor.run()