+++ tests/test_visitors/test_tokenize/test_wrong_primitives/test_unicode_prefix.py
@@ -10,42 +10,49 @@ from wemake_python_styleguide.visitors.tokenize.primitives import (
 )
 
 
-@pytest.mark.parametrize('code', [
-    'x = u"text"',
-    "print(u'unicode')",
-    '"3_3" + u"5_5"',
+@pytest.mark.parametrize('primitive', [
+    'u"text"',
+    "u'unicode'",
+    'u"5_5"',
     'u""',
 ])
 def test_unicode_prefix(
     parse_tokens,
     assert_errors,
+    assert_error_text,
     default_options,
-    code,
+    primitives_usages,
+    primitive,
+    mode,
 ):
     """Ensures that unicode prefixes raise a warning."""
-    file_tokens = parse_tokens(code)
+    file_tokens = parse_tokens(mode(primitives_usages.format(primitive)))
 
     visitor = WrongPrimitivesVisitor(default_options, file_tokens=file_tokens)
     visitor.run()
 
     assert_errors(visitor, [UnicodeStringViolation])
+    assert_error_text(visitor, primitive)
 
 
-@pytest.mark.parametrize('code', [
-    'x = "name"',
-    'x = r"text"',
-    "print(b'unicode')",
-    '"u" + "12"',
+@pytest.mark.parametrize('primitive', [
+    '"name"',
+    'r"text"',
+    "b'unicode'",
+    '"u"',
+    '"12"',
     'b""',
 ])
 def test_correct_strings(
     parse_tokens,
     assert_errors,
     default_options,
-    code,
+    primitives_usages,
+    primitive,
+    mode,
 ):
     """Ensures that correct strings are fine."""
-    file_tokens = parse_tokens(code)
+    file_tokens = parse_tokens(mode(primitives_usages.format(primitive)))
 
     visitor = WrongPrimitivesVisitor(default_options, file_tokens=file_tokens)
     visitor.run()